+++ src/Components/SourceItem/index.tsx
@@ -1,15 +1,18 @@
 import React from "react";
 import { Button } from "react-bootstrap";
 import moment from "moment";
-import { FiTrash } from "react-icons/fi";
+import { FiTrash, FiClock } from "react-icons/fi";
+import { useHistory } from "react-router-dom";
 
-import { apiCrawlers } from "../../services/api";
+import { apiCrawlers } from "../../services/apiCrawlers";
 
 import {
   Container,
   ResultDate,
   SourceName,
   TitleDateContainer,
+  ButtonStyle,
+  ButtonContainer,
 } from "./styles";
 
 export interface SourceResult {
@@ -24,9 +27,7 @@ export interface SourceResult {
   contains_end_path_keyword: boolean;
   allowed_domains: Array<string>;
   allowed_paths: Array<string>;
-  retries: number;
   page_load_timeout: number;
-  contains_dynamic_js_load: boolean;
   cron_minute: string;
   cron_hour: string;
   cron_day_of_week: string;
@@ -42,27 +43,55 @@ interface SourceItemProps {
 }
 
 const SourceItem: React.FC<SourceItemProps> = ({ item, onDelete, onClick }) => {
-  const Source = async () => {
-    await apiCrawlers
-      .delete(`/crawlers/${item.id}/`)
-      .then(() => {
-        onDelete();
-      })
-      .catch(() => {
-        alert("Ocorreu um erro inesperado ao deletar expressão!");
-      });
+  
+  const history = useHistory(); 
+  
+  const deleteSource = async (e: React.SyntheticEvent) => {
+    e.stopPropagation();
+    if (confirm('Tem certeza que deseja deletar essa fonte?')) {
+      await apiCrawlers
+        .delete(`api/crawlers/${item.id}/`)
+        .then(() => {
+          onDelete();
+        })
+        .catch(() => {
+          alert("Ocorreu um erro inesperado ao deletar a fonte! Tente novamente.");
+        });
+    }
   };
+
+  const monitoringSource = (e: React.SyntheticEvent) => {
+    e.stopPropagation();
+    history.push(`/fontes/${item.id}/monitoramento`); 
+  };
+
+  const textMargin = {
+    marginTop: '2%',
+    marginBottom: '2%'
+  }
 
   return (
     <Container onClick={() => onClick(item)} style={{ cursor: "pointer" }}>
       <TitleDateContainer>
-        <SourceName>{item.site_name}</SourceName>
+        <SourceName>{item.site_name_display}</SourceName>
         <ResultDate>
-          {moment(item.created_at).format("DD/MM/YYYY hh:mm")}
+          <ul>
+            <li style={textMargin}>Criado em: </li>
+            <li>{moment(item.created_at).format("DD/MM/YYYY hh:mm")}</li>
+          </ul>
         </ResultDate>
-        <Button onClick={Source}>
-          <FiTrash />
-        </Button>
+        <ButtonContainer>
+          <ButtonStyle>
+            <Button onClick={monitoringSource}>
+              <FiClock />
+            </Button>
+          </ButtonStyle>
+          <ButtonStyle>
+            <Button onClick={deleteSource}>
+              <FiTrash color="#ff0000"/>
+            </Button>
+          </ButtonStyle>
+        </ButtonContainer>
       </TitleDateContainer>
     </Container>
   );
